user cannot edit or delete accepted or declined request

+++ frontend/app/src/components/userRequests/Request.js
@@ -1,4 +1,4 @@
-import {FaRegEdit, FaTimes} from "react-icons/fa";
+import {FaRegEdit, FaTrashAlt} from "react-icons/fa";
 import './request.css';
 import {Modal, ModalBody, ModalHeader, ModalTitle} from "react-bootstrap";
 import EditRequest from "../userRequests/EditRequest";
@@ -17,23 +17,37 @@ export default function Request(props) {
                 <i>{props.requestRegistrationDate}</i></span>
                 <h3 className=''>
                     {props.requestId}: {props.requestType}
-
                 </h3>
+                <span><i>Status:</i></span>
+                <p className='comment'>{props.requestStatus}</p>
                 <span><i>Comment:</i></span>
                 <p className='comment'>{props.requestComment}</p>
                 <span><i>When:</i></span>
                 <p>{props.requestStartDate} - {props.requestEndDate}</p>
+
             </div>
             <div className='request-buttons'>
                 <FaRegEdit
                     className="edit-button"
-                    onClick={handleShow}
-                />
-                <FaTimes
-                    className='delete-button'
+                    size={18}
                     onClick={() => {
-                        if (window.confirm("Do you want to delete this request?")) {
-                            props.onDelete(props.requestId)
+                        if (props.requestStatus !== "PENDING") {
+                            alert("You cannot edit accepted or declined request.")
+                        } else {
+                            handleShow();
+                        }
+                    }}
+                />
+                <FaTrashAlt
+                    className='delete-button'
+                    size={15}
+                    onClick={() => {
+                        if (props.requestStatus === "PENDING") {
+                            if (window.confirm("Do you want to delete this request?")) {
+                                props.onDelete(props.requestId)
+                            }
+                        } else {
+                            alert("You cannot delete accepted or declined request.")
                         }
                     }
                     }
